DRY for limit trait tests, actualize HasLength
